MP-149 :truck: Rename: 팁 좋아요 폴더 구조 변경 * 기존 : tip 디렉토리 내 like 폴더 * 변경 : tip 디렉토리 폴더로 이동
