+++ src/main/java/kr/modusplant/domains/communication/tip/app/http/response/TipLikeResponse.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.communication.tip.app.http.like.response;
+package kr.modusplant.domains.communication.tip.app.http.response;
 
 public record TipLikeResponse(int likeCount, boolean liked) {
     public static TipLikeResponse of(int likeCount, boolean liked) {
