+++ src/main/java/kr/modusplant/domains/communication/tip/app/service/TipLikeApplicationService.java
@@ -1,10 +1,10 @@
-package kr.modusplant.domains.communication.tip.app.service.like;
+package kr.modusplant.domains.communication.tip.app.service;
 
-import kr.modusplant.domains.communication.tip.app.http.like.response.TipLikeResponse;
-import kr.modusplant.domains.communication.tip.domain.service.like.TipLikeValidationService;
+import kr.modusplant.domains.communication.tip.app.http.response.TipLikeResponse;
+import kr.modusplant.domains.communication.tip.domain.service.TipLikeValidationService;
 import kr.modusplant.domains.communication.tip.persistence.entity.TipPostEntity;
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeEntity;
-import kr.modusplant.domains.communication.tip.persistence.like.repository.TipLikeRepository;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeEntity;
+import kr.modusplant.domains.communication.tip.persistence.repository.TipLikeRepository;
 import kr.modusplant.domains.communication.tip.persistence.repository.TipPostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
