+++ src/main/java/kr/modusplant/domains/communication/tip/domain/service/TipLikeValidationService.java
@@ -1,10 +1,10 @@
-package kr.modusplant.domains.communication.tip.domain.service.like;
+package kr.modusplant.domains.communication.tip.domain.service;
 
 import kr.modusplant.domains.communication.tip.persistence.entity.TipPostEntity;
 import kr.modusplant.domains.communication.tip.persistence.repository.TipPostRepository;
 import kr.modusplant.domains.member.persistence.entity.SiteMemberEntity;
 import kr.modusplant.domains.member.persistence.repository.SiteMemberRepository;
-import kr.modusplant.domains.communication.tip.persistence.like.repository.TipLikeRepository;
+import kr.modusplant.domains.communication.tip.persistence.repository.TipLikeRepository;
 import kr.modusplant.global.error.EntityExistsWithUuidException;
 import kr.modusplant.global.error.EntityNotFoundWithUlidException;
 import lombok.RequiredArgsConstructor;
