+++ src/main/java/kr/modusplant/domains/communication/tip/persistence/repository/TipLikeRepository.java
@@ -1,7 +1,7 @@
-package kr.modusplant.domains.communication.tip.persistence.like.repository;
+package kr.modusplant.domains.communication.tip.persistence.repository;
 
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeEntity;
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeId;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeEntity;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
