+++ src/test/java/kr/modusplant/domains/communication/tip/app/service/TipLikeApplicationServiceTest.java
@@ -1,11 +1,11 @@
-package kr.modusplant.domains.communication.tip.app.service.like;
+package kr.modusplant.domains.communication.tip.app.service;
 
-import kr.modusplant.domains.communication.tip.app.http.like.response.TipLikeResponse;
+import kr.modusplant.domains.communication.tip.app.http.response.TipLikeResponse;
 import kr.modusplant.domains.communication.tip.common.util.entity.TipPostEntityTestUtils;
 import kr.modusplant.domains.communication.tip.persistence.entity.TipPostEntity;
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeEntity;
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeId;
-import kr.modusplant.domains.communication.tip.persistence.like.repository.TipLikeRepository;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeEntity;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeId;
+import kr.modusplant.domains.communication.tip.persistence.repository.TipLikeRepository;
 import kr.modusplant.domains.communication.tip.persistence.repository.TipPostRepository;
 import kr.modusplant.domains.member.common.util.entity.SiteMemberEntityTestUtils;
 import kr.modusplant.domains.member.persistence.entity.SiteMemberEntity;
