+++ src/test/java/kr/modusplant/domains/communication/tip/common/util/domain/TipLikeTestUtils.java
@@ -1,7 +1,6 @@
-package kr.modusplant.domains.communication.tip.common.util.like.domain;
+package kr.modusplant.domains.communication.tip.common.util.domain;
 
-import kr.modusplant.domains.communication.tip.common.util.domain.TipPostTestUtils;
-import kr.modusplant.domains.communication.tip.domain.model.like.TipLike;
+import kr.modusplant.domains.communication.tip.domain.model.TipLike;
 import kr.modusplant.domains.member.common.util.domain.SiteMemberTestUtils;
 
 public interface TipLikeTestUtils extends TipPostTestUtils, SiteMemberTestUtils {
