+++ src/test/java/kr/modusplant/domains/communication/tip/common/util/entity/TipLikeEntityTestUtils.java
@@ -1,7 +1,6 @@
-package kr.modusplant.domains.communication.tip.common.util.like.entity;
+package kr.modusplant.domains.communication.tip.common.util.entity;
 
-import kr.modusplant.domains.communication.tip.common.util.entity.TipPostEntityTestUtils;
-import kr.modusplant.domains.communication.tip.persistence.like.entity.TipLikeEntity;
+import kr.modusplant.domains.communication.tip.persistence.entity.TipLikeEntity;
 import kr.modusplant.domains.member.common.util.entity.SiteMemberEntityTestUtils;
 
 public interface TipLikeEntityTestUtils extends TipPostEntityTestUtils, SiteMemberEntityTestUtils {
