+++ src/test/java/kr/modusplant/domains/communication/tip/domain/service/TipLikeValidationServiceTest.java
@@ -1,6 +1,6 @@
-package kr.modusplant.domains.communication.tip.domain.service.like;
+package kr.modusplant.domains.communication.tip.domain.service;
 
-import kr.modusplant.domains.communication.tip.persistence.like.repository.TipLikeRepository;
+import kr.modusplant.domains.communication.tip.persistence.repository.TipLikeRepository;
 import kr.modusplant.domains.communication.tip.persistence.repository.TipPostRepository;
 import kr.modusplant.domains.member.persistence.repository.SiteMemberRepository;
 import kr.modusplant.global.error.EntityExistsWithUuidException;
